+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Changed the error numbers for the user errors to avoid overlap.
 - Fixed a bug in the long/lat geocode function.
+- Added a generic place collection class.
 
 **1.0.0.2024** *(May 24, 2018)*